Fix compiler warnings and build error across five source files

+++ assemble.c
@@ -1039,7 +1039,7 @@ static void PassOneChecks( void )
          */
 #if FASTPASS
         /* UASM 2.55 , altname is set to 1 when the type is undefined */
-        if ( curr->sym.altname > 1 ) 
+        if ( (size_t)curr->sym.altname > 1 ) 
 		{
             if ( curr->sym.altname->state == SYM_INTERNAL ) 
 			{
